--- MA_PROYECTO/.idea/dictionaries/cesar.xml
@@ -1,3 +1,0 @@
-<component name="ProjectDictionaryState">
-  <dictionary name="cesar" />
-</component>
