--- MA_PROYECTO/.idea/externalDependencies.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="ExternalDependencies">
-    <plugin id="com.intellij.java" />
-  </component>
-</project>
