--- MA_PROYECTO/.idea/project-template.xml
@@ -1,3 +1,0 @@
-<template>
-  <input-field default="com.company">IJ_BASE_PACKAGE</input-field>
-</template>
