+++ MA_PROYECTO/src/test/java/ec/edu/epn/proyecto/MenuTest.java
@@ -1,0 +1,7 @@
+package ec.edu.epn.proyecto;
+
+import static org.junit.Assert.*;
+
+public class MenuTest {
+
+}
